% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{genr_sim_data}
\alias{genr_sim_data}
\title{Simulate Data from Dube et al. (2023)}
\usage{
genr_sim_data(exogenous_timing = TRUE, seed = 757)
}
\arguments{
\item{exogenous_timing}{A boolean value that, if set to TRUE, will make treatment exogenous.  When set to FALSE,
previous values of Y will make treatment more likely, or endogenous.}

\item{seed}{A numeric value to set.seed().}
}
\value{
A simulated dataset.
}
\description{
This function generates simulated data from Dube et al. (2023).
}
