% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_speed.R
\name{lpdid_dt_reg}
\alias{lpdid_dt_reg}
\title{Local Projections Difference-in-Differences Regression}
\usage{
lpdid_dt_reg(
  df,
  unit_index,
  time_index,
  cluster = NULL,
  controls = NULL,
  controls_t = NULL,
  reweight = FALSE
)
}
\arguments{
\item{df}{The dataset used in the analysis.}

\item{unit_index}{The name of the column that represents the unit ID.  This should be a character.}

\item{time_index}{The name of the column that represents the calendar time.  This should be a character.}

\item{cluster}{The name of the column by which to cluster the standard errors.  Default is the unit ID.}

\item{controls}{A formula of control variables to be included in the regression formula. The form should be: `~ X1 + X2 | FE1 + FE2`. These controls will be time invariant.}

\item{controls_t}{A formula of control variables to be included in the regression formula. The form should be: `~ X1 + X2 | FE1 + FE2`. These controls will be time varying.}

\item{reweight}{A boolean (TRUE or FALSE) value that will re-weight the regression and generate ATT rather than VWATT. The default is FALSE, which corresponds to the estimator calculating the VWATT (variance weighted average treatment effect on the treated).}

\item{outcome_lags}{The number of outcome lags to be included in the analysis.
For an example of this, simulate endogenous data via genr_sim_data(FALSE), and include one outcome lag.}
}
\value{
A list including a coefficient table and the data.table for the regression.
}
\description{
This function estimates LP-DiD regressions as outlined in Dube et al. (2023) <doi:10.3386/w31184>.
}
