% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{lpdid}
\alias{lpdid}
\title{Local Projections Difference-in-Differences}
\usage{
lpdid(
  df,
  window = c(NA, NA),
  y,
  unit_index,
  time_index,
  rel_time = "",
  controls = NULL,
  outcome_lags = 0,
  reweight = FALSE,
  pmd = FALSE,
  pmd_lag,
  composition_correction = FALSE,
  pooled = FALSE,
  nonabsorbing = FALSE,
  nonabsorbing_lag,
  nonabsorbing_treat_status = ""
)
}
\arguments{
\item{df}{The dataset used in the analysis.}

\item{window}{A vector of length two that denotes the length of the pre- and post-periods.
The first number should be negative and second should be positive.}

\item{y}{The outcome variable.  This should be a character.}

\item{unit_index}{The name of the column that represents the unit ID.  This should be a character.}

\item{time_index}{The name of the column that represents the calendar time.  This should be a character.}

\item{rel_time}{The name of the column that contains a "time to treatment" vector.
Values for untreated units must all be negative.}

\item{controls}{A vector of names of control variables to be included in the regression formula.}

\item{outcome_lags}{The number of outcome lags to be included in the analysis.
For an example of this, simulate endogenous data via genr_sim_data(FALSE), and include one outcome lag.}

\item{reweight}{A boolean value that will re-weight the regression and generate ATT rather than VWATT.}

\item{pmd}{A boolean value that, if equal to TRUE, will use pre-treatment means rather than a single value from t-1.}

\item{pmd_lag}{The number of pre-treatment periods to include in taking the mean.}

\item{composition_correction}{A boolean value that will remove later-treated observations from the control group even before they are treated.  See Section 2.10 "Composition effects".}

\item{pooled}{A boolean value that, if equal to TRUE, will calculate the treatment effect pooled over all post-periods.}

\item{nonabsorbing}{A boolean value that, if equal to TRUE, will preform a routine similar to "stacking".
Using this option requires a different variable than "rel_time".  More details can be seen in "nonabsorbing_treatment_status".}

\item{nonabsorbing_lag}{Sets the number of periods after which dynamic effects stabilize.}

\item{nonabsorbing_treat_status}{The name of the column that denotes treatment status.
This vector must take on a value of 1 when the unit is treated and dynamics are still in play and zero otherwise.
As an example, in a state-year panel, if a state is treated in 1990 and 2010, and dynamics settle after 5 years, the state should have 1's for 1990-1994 and 2010-2014.}
}
\value{
A list including a coefficient table and window vector.
}
\description{
This function estimates LP-DiD regressions as outlined in Dube et al. (2023) <doi:10.3386/w31184>.
}
