% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_lpdid}
\alias{plot_lpdid}
\title{Plot LP-DiD Event Study}
\usage{
plot_lpdid(
  reg,
  conf = 0.95,
  segments = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = "",
  pch = 19,
  cex = 1,
  col = "black",
  opacity = 1
)
}
\arguments{
\item{reg}{An object generated via the lpdid function.}

\item{conf}{The confidence level (1-alpha) desired for confidence intervals.}

\item{segments}{A boolean value that will generate confidence intervals.  Default is TRUE.}
}
\value{
An event study plot.
}
\description{
Plotting LP-DiD event study parameter estimates and confidence intervals.
}
